% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_rainy.R
\name{count_rainy}
\alias{count_rainy}
\title{Count rainy days in the weather data.}
\usage{
count_rainy(.data, rf = 10)
}
\arguments{
\item{.data}{The weather data contains precipitation.}

\item{rf}{The least amount of precipitation needed to be a rainy day.}
}
\value{
Integer, number of rainy days.
}
\description{
Count rainy days in the weather data.
}
