% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincouver_core.R
\name{format_url}
\alias{format_url}
\title{Form a url with parameters.}
\usage{
format_url(url, params)
}
\arguments{
\item{url}{The base url you want to add parameters to.}

\item{params}{The url parameters you want to add.}
}
\value{
`string` contains formatted url.
}
\description{
Form a url with parameters.
}
