% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download daily weather data and convert to `data.frame`.}
\usage{
get_data(y = 2018)
}
\arguments{
\item{y}{The year of the data you want.}
}
\value{
`data.frame` contains weather data, including temperature and precipitation.
}
\description{
Download daily weather data and convert to `data.frame`.
}
