% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate.R
\name{plot_climate}
\alias{plot_climate}
\title{Plot the temperature range and preticipation amount.}
\usage{
plot_climate(.data)
}
\arguments{
\item{.data}{The weather data contains temperature and precipitation.}
}
\value{
This function returns nothing.
}
\description{
Plot the temperature range and preticipation amount.
}
