% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincouver_core.R
\name{preprocess_xml}
\alias{preprocess_xml}
\title{Preprocess the xml file, select the data nodes only.}
\usage{
preprocess_xml(xml)
}
\arguments{
\item{xml}{The xml file contains the weather data.}
}
\value{
`xmlNode` contains weather data in xml format.
}
\description{
Preprocess the xml file, select the data nodes only.
}
