% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source.R
\docType{data}
\name{url_base}
\alias{url_base}
\title{The url from where the weather is downloaded.}
\format{An object of class \code{character} of length 1.}
\usage{
url_base
}
\description{
The url from where the weather is downloaded.
}
\keyword{datasets}
