% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source.R
\docType{data}
\name{url_params}
\alias{url_params}
\title{The url parameters used in downloading.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.}
\usage{
url_params
}
\description{
The url parameters used in downloading.
}
\keyword{datasets}
