% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmltodf.R
\name{xmltodf}
\alias{xmltodf}
\title{Clean and tidy the xml data file, convert to `data.frame`.}
\usage{
xmltodf(xml)
}
\arguments{
\item{xml}{The xml data file which contains the weather data only.}
}
\value{
`data.frame` contains tidy weather data.
}
\description{
Clean and tidy the xml data file, convert to `data.frame`.
}
